tehtava 4.21

+++ part4/controllers/blogs.js
@@ -8,6 +8,15 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
+blogsRouter.get('/:id', async(request, response) => {
+    const blog = await Blog.findById(request.params.id)
+    if (blog) {
+        response.json(blog.toJSON())
+    } else {
+        response.status(404).end()
+    }
+})
+
 /*const getTokenFrom = request => {
     const authorization = request.get('authorization')
     if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
@@ -50,12 +59,29 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    try {
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({error: 'token missing or invalid'})
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({error: 'not authorized'})
+    }
+
+    /*try {
         await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
     } catch (exception) {
         next(exception)
-    }
+    }*/
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
